# bthwani-backend-nest/reports/openapi.yaml
openapi: 3.0.0
paths:
...
  /cart:
    get:
      operationId: CartController_getMyCart
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: الحصول على سلتي
      tags:
        - Cart
    delete:
      operationId: CartController_clearCart
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تفريغ السلة
      tags:
        - Cart
  /cart/items:
    post:
      operationId: CartController_addToCart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إضافة منتج للسلة
      tags:
        - Cart
  /cart/items/{productId}:
    patch:
      operationId: CartController_updateCartItem
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث كمية منتج
      tags:
        - Cart
    delete:
      operationId: CartController_removeFromCart
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف منتج من السلة
      tags:
        - Cart
  /cart/note:
    patch:
      operationId: CartController_addNote
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddNoteDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إضافة ملاحظة
      tags:
        - Cart
...
  /cart/count:
    get:
      operationId: CartController_getCartCount
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: عدد العناصر في السلة
      tags:
        - Cart
  /cart/shein:
    get:
      operationId: CartController_getMySheinCart
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: الحصول على سلة Shein
      tags:
        - Cart
    delete:
      operationId: CartController_clearSheinCart
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تفريغ سلة Shein
      tags:
        - Cart
  /cart/shein/items:
    post:
      operationId: CartController_addToSheinCart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToSheinCartDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إضافة منتج Shein للسلة
      tags:
        - Cart
  /cart/shein/items/{sheinProductId}:
    patch:
      operationId: CartController_updateSheinCartItem
      parameters:
        - name: sheinProductId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSheinCartItemDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث كمية منتج Shein
      tags:
        - Cart
    delete:
      operationId: CartController_removeFromSheinCart
      parameters:
        - name: sheinProductId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف منتج Shein من السلة
      tags:
        - Cart
...
  /cart/shein/note:
    patch:
      operationId: CartController_addSheinNote
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddNoteDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إضافة ملاحظة لسلة Shein
      tags:
        - Cart
  /cart/combined:
    get:
      operationId: CartController_getCombinedCart
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: الحصول على السلة الموحدة
      tags:
        - Cart
  /cart/combined/clear-all:
    delete:
      operationId: CartController_clearAllCarts
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تفريغ كل السلات
      tags:
        - Cart
...
info:
  title: Bthwani API v2
  description: NestJS API Documentation - نظام إدارة الطلبات والتجارة
  version: '2.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AddToCartDto:
      type: object
      properties: {}
    UpdateCartItemDto:
      type: object
      properties: {}
    AddNoteDto:
      type: object
      properties: {}
...
    AddToSheinCartDto:
      type: object
      properties: {}
    UpdateSheinCartItemDto:
      type: object
      properties: {}
...